+++ src/webOAuthServer.ts
@@ -11,6 +11,7 @@ import * as http from 'http';
 import { parse as parseQueryString } from 'querystring';
 import { parse as parseUrl } from 'url';
 import { Socket } from 'net';
+import { EventEmitter } from 'events';
 import { JwtOAuth2Config, OAuth2 } from 'jsforce';
 import { AsyncCreatable, Env, set, toNumber } from '@salesforce/kit';
 import { asString, get, Nullable } from '@salesforce/ts-types';
@@ -46,6 +47,7 @@ export class WebOAuthServer extends AsyncCreatable<WebOAuthServer.Options> {
   private webServer!: WebServer;
   private oauth2!: OAuth2;
   private oauthConfig: JwtOAuth2Config;
+  private oauthError = new Error('Oauth Error');
 
   public constructor(options: WebOAuthServer.Options) {
     super(options);
@@ -94,11 +96,12 @@ export class WebOAuthServer extends AsyncCreatable<WebOAuthServer.Options> {
                 oauth2: this.oauth2,
               });
               await authInfo.save();
-              this.webServer.doRedirect(303, authInfo.getOrgFrontDoorUrl(), response);
+              await this.webServer.handleSuccess(response);
               response.end();
               resolve(authInfo);
             } catch (err) {
-              this.webServer.reportError(err as Error, response);
+              this.oauthError = err as Error;
+              await this.webServer.handleError(response);
               reject(err);
             }
           })
@@ -148,40 +151,52 @@ export class WebOAuthServer extends AsyncCreatable<WebOAuthServer.Options> {
   private async executeOauthRequest(): Promise<http.ServerResponse> {
     return new Promise((resolve, reject) => {
       this.logger.debug('Starting web auth flow');
-      // eslint-disable-next-line @typescript-eslint/no-misused-promises, @typescript-eslint/no-explicit-any, @typescript-eslint/require-await
-      this.webServer.server.on('request', async (request: any, response) => {
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-        const url = parseUrl(request.url);
-        this.logger.debug(`processing request for uri: ${url.pathname as string}`);
-        if (request.method === 'GET') {
-          if (url.pathname?.startsWith('/OauthRedirect')) {
-            request.query = parseQueryString(url.query as string);
-            if (request.query.error) {
-              // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-              const err = new SfError(request.query.error_description ?? request.query.error, request.query.error);
-              this.webServer.reportError(err, response);
-              return reject(err);
-            }
-            this.logger.debug(`request.query.state: ${request.query.state as string}`);
-            try {
-              // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-              this.oauthConfig.authCode = asString(this.parseAuthCodeFromRequest(response, request));
-              resolve(response);
-            } catch (err) {
-              reject(err);
+      // - async method when sync expected
+      // eslint-disable-next-line @typescript-eslint/no-misused-promises
+      this.webServer.server.on('request', async (request: WebOAuthServer.Request, response) => {
+        if (request.url) {
+          const url = parseUrl(request.url);
+          this.logger.debug(`processing request for uri: ${url.pathname}`);
+          if (request.method === 'GET') {
+            if (url.pathname?.startsWith('/OauthRedirect') && url.query) {
+              request.query = parseQueryString(url.query) as {
+                code: string;
+                state: string;
+                error?: string | undefined;
+                error_description?: string;
+              };
+              if (request.query.error) {
+                const errorName: string =
+                  typeof request.query.error_description === 'string'
+                    ? request.query.error_description
+                    : request.query.error;
+                this.oauthError = new SfError(errorName, request.query.error);
+                await this.webServer.handleError(response);
+                return reject(this.oauthError);
+              }
+              this.logger.debug(`request.query.state: ${request.query.state}`);
+              try {
+                this.oauthConfig.authCode = asString(this.parseAuthCodeFromRequest(response, request));
+                resolve(response);
+              } catch (err) {
+                reject(err);
+              }
+            } else if (url.pathname === '/OauthSuccess') {
+              this.webServer.reportSuccess(response);
+            } else if (url.pathname === '/OauthError') {
+              this.webServer.reportError(this.oauthError, response);
+            } else {
+              this.webServer.sendError(404, 'Resource not found', response);
+              const errName = 'invalidRequestUri';
+              const errMessage = messages.getMessage(errName, [url.pathname]);
+              reject(new SfError(errMessage, errName));
             }
           } else {
-            this.webServer.sendError(404, 'Resource not found', response);
-            const errName = 'invalidRequestUri';
-            const errMessage = messages.getMessage(errName, [url.pathname]);
+            this.webServer.sendError(405, 'Unsupported http methods', response);
+            const errName = 'invalidRequestMethod';
+            const errMessage = messages.getMessage(errName, [request.method]);
             reject(new SfError(errMessage, errName));
           }
-        } else {
-          this.webServer.sendError(405, 'Unsupported http methods', response);
-          const errName = 'invalidRequestMethod';
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-          const errMessage = messages.getMessage(errName, [request.method]);
-          reject(new SfError(errMessage, errName));
         }
       });
     });
@@ -249,7 +264,9 @@ export namespace WebOAuthServer {
     oauthConfig: JwtOAuth2Config;
   }
 
-  export type Request = http.IncomingMessage & { query: { code: string; state: string } };
+  export type Request = http.IncomingMessage & {
+    query: { code: string; state: string; error?: string; error_description?: string };
+  };
 }
 
 /**
@@ -262,6 +279,7 @@ export class WebServer extends AsyncCreatable<WebServer.Options> {
   public host = 'localhost';
   private logger!: Logger;
   private sockets: Socket[] = [];
+  private redirectStatus = new EventEmitter();
 
   public constructor(options: WebServer.Options) {
     super(options);
@@ -312,7 +330,7 @@ export class WebServer extends AsyncCreatable<WebServer.Options> {
   /**
    * sends a response error.
    *
-   * @param statusCode he statusCode for the response.
+   * @param status the statusCode for the response.
    * @param message the message for the http body.
    * @param response the response to write the error to.
    */
@@ -325,11 +343,12 @@ export class WebServer extends AsyncCreatable<WebServer.Options> {
   /**
    * sends a response redirect.
    *
-   * @param statusCode the statusCode for the response.
+   * @param status the statusCode for the response.
    * @param url the url to redirect to.
    * @param response the response to write the redirect to.
    */
   public doRedirect(status: number, url: string, response: http.ServerResponse): void {
+    this.logger.debug(`Redirecting to ${url}`);
     response.setHeader('Content-Type', 'text/plain');
     const body = `${status} - Redirecting to ${url}`;
     response.setHeader('Content-Length', Buffer.byteLength(body));
@@ -340,18 +359,57 @@ export class WebServer extends AsyncCreatable<WebServer.Options> {
   /**
    * sends a response to the browser reporting an error.
    *
-   * @param error the error
-   * @param response the response to write the redirect to.
+   * @param error the oauth error
+   * @param response the HTTP response.
    */
   public reportError(error: Error, response: http.ServerResponse): void {
     response.setHeader('Content-Type', 'text/html');
-    const body = messages.getMessage('serverErrorHTMLResponse', [error.message]);
-    response.setHeader('Content-Length', Buffer.byteLength(body));
+    const currentYear = new Date().getFullYear();
+    const encodedImg = messages.getMessage('serverSfdcImage');
+    const body = messages.getMessage('serverErrorHTMLResponse', [encodedImg, error.name, error.message, currentYear]);
+    response.setHeader('Content-Length', Buffer.byteLength(body, 'utf8'));
     response.end(body);
+    if (error.stack) {
+      this.logger.debug(error.stack);
+    }
+    this.redirectStatus.emit('complete');
+  }
+
+  /**
+   * sends a response to the browser reporting the success.
+   *
+   * @param response the HTTP response.
+   */
+  public reportSuccess(response: http.ServerResponse): void {
+    response.setHeader('Content-Type', 'text/html');
+    const currentYear = new Date().getFullYear();
+    const encodedImg = messages.getMessage('serverSfdcImage');
+    const body = messages.getMessage('serverSuccessHTMLResponse', [encodedImg, currentYear]);
+    response.setHeader('Content-Length', Buffer.byteLength(body, 'utf8'));
+    response.end(body);
+    this.redirectStatus.emit('complete');
+  }
+
+  public async handleSuccess(response: http.ServerResponse): Promise<void> {
+    return this.handleRedirect(response, '/OauthSuccess');
+  }
+
+  public async handleError(response: http.ServerResponse): Promise<void> {
+    return this.handleRedirect(response, '/OauthError');
   }
 
   protected async init(): Promise<void> {
     this.logger = await Logger.child(this.constructor.name);
+  }
+
+  private async handleRedirect(response: http.ServerResponse, url: '/OauthSuccess' | '/OauthError'): Promise<void> {
+    return new Promise((resolve) => {
+      this.redirectStatus.on('complete', () => {
+        this.logger.debug('Redirect complete');
+        resolve();
+      });
+      this.doRedirect(303, url, response);
+    });
   }
 
   /**